@@ -14,30 +14,29 @@ description: A new Flutter project.
 version: 1.0.1+2
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
+  clay_containers: ^0.3.2
+  cupertino_icons: ^1.0.5
+  flip_card: ^0.7.0
   flutter:
     sdk: flutter
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  flutter_flip_view: ^1.0.3
-  flutter_neumorphic: ^1.0.4
-  sweetsheet: ^0.2.0
-  flutter_keyboard_visibility: ^0.8.0
-  clay_containers: ^0.2.2
-  shared_preferences: ^0.5.6+3
+  flutter_keyboard_visibility: ^5.4.1
+  # flutter_neumorphic: ^3.2.0
+  flutter_neumorphic: 
+    git:
+      url: https://github.com/den0206/Flutter-Neumorphic.git
+      ref: feature/remoce_accentcolor
+  shared_preferences: ^2.2.0
+  sweetsheet: ^0.4.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
 
@@ -45,18 +44,14 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
-
   # To add assets to your application, add an assets section, like this:
   # assets:
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
-
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
