@@ -1,56 +1,65 @@
 import 'package:clay_containers/clay_containers.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
+import 'package:spongemock_flutter/utils/app_theme.dart';
 
 class NAppBar extends StatefulWidget {
   final String title;
-  final ValueChanged<bool> onThemeToggleChanged;
+  final ValueChanged<bool>? onThemeToggleChanged;
   final bool themeToggleValue;
   NAppBar({
-    @required this.title,
+    required this.title,
     this.onThemeToggleChanged,
     this.themeToggleValue = false,
-  }) : assert(title != null);
+  });
   @override
   _NAppBarState createState() => _NAppBarState();
 }
 
 class _NAppBarState extends State<NAppBar> {
+  late final AppTheme _theme;
+
+  @override
+  void initState() {
+    _theme = AppTheme(context);
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
+    final isDarkTheme = _theme.isDarkTheme;
     var appBar = AppBar(
       centerTitle: true,
       backgroundColor: Colors.transparent,
       elevation: 0,
+
       /// wrap in [AnimatedContainer] and match duration to settings of
-      /// [NeumorphicBackground] to maintain a consistent transition 
+      /// [NeumorphicBackground] to maintain a consistent transition
       /// between themes (See [NeumorphicBackground] source)
       title: AnimatedContainer(
-        color: NeumorphicTheme.of(context).current.baseColor,
+        color: NeumorphicTheme.of(context)?.current?.baseColor,
         duration: const Duration(milliseconds: 100),
         child: ClayText(
           widget.title,
           size: 30,
           emboss: true,
-          color: NeumorphicTheme.of(context).isUsingDark
+          color: isDarkTheme
               ? Colors.grey[700]
-              : NeumorphicTheme.of(context).current.baseColor,
+              : _theme.neomorphTheme?.current?.baseColor,
         ),
       ),
       actions: <Widget>[
         Row(
           children: <Widget>[
-            NeumorphicTheme.of(context).isUsingDark
-                ? Container()
-                : Icon(
-                    Icons.wb_sunny,
-                    color: NeumorphicTheme.of(context).current.defaultTextColor,
-                  ),
+            if (!isDarkTheme)
+              Icon(
+                Icons.wb_sunny,
+                color: _theme.neomorphTheme?.current?.defaultTextColor,
+              ),
             Padding(
               padding: const EdgeInsets.all(14.0),
               child: NeumorphicSwitch(
                 onChanged: (value) {
-                  widget.onThemeToggleChanged(value);
+                  widget.onThemeToggleChanged?.call(value);
                 },
                 value: widget.themeToggleValue,
                 style: NeumorphicSwitchStyle(
@@ -58,19 +67,16 @@ class _NAppBarState extends State<NAppBar> {
                 ),
               ),
             ),
-            NeumorphicTheme.of(context).isUsingDark
-                ? Row(
-                    children: <Widget>[
-                      Icon(
-                        Icons.brightness_3,
-                        color: NeumorphicTheme.of(context)
-                            .current
-                            .defaultTextColor,
-                      ),
-                      SizedBox(width: 8),
-                    ],
-                  )
-                : Container(),
+            if (isDarkTheme)
+              Row(
+                children: <Widget>[
+                  Icon(
+                    Icons.brightness_3,
+                    color: _theme.neomorphTheme?.current?.defaultTextColor,
+                  ),
+                  SizedBox(width: 8),
+                ],
+              )
           ],
         ),
       ],
